backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Environment Variables
python-decouple==3.8

# CORS Headers
django-cors-headers==4.3.1

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2

# Data Validation
validators==0.22.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
